Follow-up to r1405530:
* subversion/svn/mergeinfo-cmd.c
  (mergeinfo_diagram): silence yet another conversion warning

git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1405552 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svn/mergeinfo-cmd.c
@@ -152,7 +152,7 @@ mergeinfo_diagram(svn_client__pathrev_t *yca,
       col_width[col] = 0;
       for (row = 0; row < ROWS; row++)
         {
-          if (g[row][col] && (strlen(g[row][col]) > col_width[col]))
+          if (g[row][col] && ((int)strlen(g[row][col]) > col_width[col]))
             col_width[col] = (int)strlen(g[row][col]);
         }
     }
